pyproject.toml
[tool.poetry]
name = "project-office"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django = "4.2.0"
drf-spectacular = "^0.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
